add client id settings
This adds a new config setting `clientid` in the server settings:
When unset by the user, it generates a unique one in the form of
"gotify-012345678abcdef".

The PR also fixes that the plugin does not connect on startup of
gotify (or after enabling it). Prior to this a config change was
required to connect to the broker...